# MultiAPI/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Multi_TODO_API",
    "version": "v1"
  },
  "paths": {
    "/api/Authenticate/login": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Authenticate/register": {
      "post": {
        "tags": [
          "Authenticate"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TodoItem": {
      "get": {
        "tags": [
          "TodoItem"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TodoItem"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TodoItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TodoItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TodoItem"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TodoItem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TodoItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TodoItem"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TodoItem/{id}": {
      "get": {
        "tags": [
          "TodoItem"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TodoItem"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Login": {
      "required": [
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Register": {
      "required": [
        "email",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "format": "email",
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "TodoItem": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "accountId": {
          "format": "int32",
          "type": "integer"
        },
        "createdAt": {
          "type": "string"
        },
        "modifiedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "modifyBy": {
          "type": "string"
        }
      }
    }
  }
}